+++ src/elements/map-control-simple-layer-list/config.js
@@ -1,0 +1,2 @@
+export const elementName = "map-control-simple-layer-list";
+export const olControlElementName = "map-control-simple-layer-list--ol-control";