+++ src/elements/map-layer-geojson/index.js
@@ -5,7 +5,10 @@ import {
 
 import HTMLMapLayerBase from '../map-layer-base';
 
-const defaultDataProjection = 'EPSG:4326';
+import {
+  elementName,
+  defaultDataProjection,
+} from './config';
 
 /**
  * Usage:
@@ -19,7 +22,7 @@ const defaultDataProjection = 'EPSG:4326';
  *   // Specify the projection the source data coordinates are in. It will only be used when no CRS is available in the data. Default value is "EPSG:4326".
  *   src-projection="{string}"
  * >
- *   <HTMLMapLayerVectorStyle ... />
+ *   <HTMLMapLayerVectorStyle ...></HTMLMapLayerVectorStyle>
  * </HTMLMapLayerGeoJSON>
  */
 export default class HTMLMapLayerGeoJSON extends HTMLMapLayerBase {
@@ -244,3 +247,5 @@ export default class HTMLMapLayerGeoJSON extends HTMLMapLayerBase {
    */
 
 } // HTMLMapLayerGeoJSON
+
+customElements.define(elementName, HTMLMapLayerGeoJSON);