+++ src/elements/map-layer-group/index.js
@@ -1,10 +1,14 @@
 import HTMLMapLayerBase from '../map-layer-base';
 
+import {
+  elementName,
+} from './config';
+
 /**
  * Usage:
  * <HTMLMapLayerGroup
  *   // @inheritdoc
- * />
+ * ></HTMLMapLayerGroup>
  */
 export default class HTMLMapLayerGroup extends HTMLMapLayerBase {
 
@@ -33,17 +37,35 @@ export default class HTMLMapLayerGroup extends HTMLMapLayerBase {
   }
 
   /**
-   * A static helper function for setting up observers for monitoring child layer element changes.
+   * Returns a collection of the child layer elements.
+   * The provided layer collection will also be auto-populated with all the layers.
+   * The collection is auto-updated until `.disconnect()` is called.
    * Does it potentially leak observers?
    * @param {HTMLElement} element
    * @param {ol.Collection.<ol.layer.Base>} [layerCollection]
    * @returns {ol.Collection.<HTMLMapLayerBase>}
    */
-  static setupChildLayerElementsObserver (element, layerCollection) {
-    const elementCollection = this.setupChildElementsObserver(element, HTMLMapLayerBase);
+  static getLiveChildLayerElementCollection (element, layerCollection) {
+    const elementCollection = this.getLiveChildElementCollection(element, HTMLMapLayerBase);
 
-    elementCollection.on('change', ({/*type, */target}) => {
-      const layerElements = target.getArray();
+    // Cache loaded items for comparison when items are updated.
+    elementCollection._cachedItems = new Set();
+    elementCollection._childLayerGroupOnChange = (function () {
+      // When any child layer group has been changed, this layer group is also changed.
+      this.changed();
+    }).bind(elementCollection);
+
+    elementCollection.on('change', ({/*type, */target: collection}) => {
+      const layerElements = collection.getArray();
+
+      // Find out what's added and what's removed.
+      const layerElementSet = new Set(layerElements);
+      const addedElements = layerElements.filter((el) => !elementCollection._cachedItems.has(el));
+      const removedElements = Array.from(elementCollection._cachedItems).filter((el) => !layerElementSet.has(el));
+
+      // Update the cache.
+      elementCollection._cachedItems.clear();
+      elementCollection._cachedItems = new Set(layerElements);
 
       // Children should have the same projection as the parent.
       this.alignLayerElementProjections(element, layerElements);
@@ -56,25 +78,42 @@ export default class HTMLMapLayerGroup extends HTMLMapLayerBase {
         layerCollection.extend(layers);
         layerCollection.changed();
       }
+
+      // Remove listeners from removed elements.
+      removedElements
+        .filter((el) => el instanceof HTMLMapLayerGroup)
+        .forEach((el) => el.removeEventListener('change:layers', elementCollection._childLayerGroupOnChange));
+
+      // Add listeners to added elements.
+      addedElements
+        .filter((el) => el instanceof HTMLMapLayerGroup)
+        .forEach((el) => el.addEventListener('change:layers', elementCollection._childLayerGroupOnChange));
+
+      element.dispatchEvent(new Event('change:layers'));
     });
 
     return elementCollection;
   }
 
-  /**
-   * An instance of the element is created or upgraded. Useful for initializing state, settings up event listeners, or creating shadow dom. See the spec for restrictions on what you can do in the constructor.
-   */
   constructor () {
-    super(); // always call super() first in the ctor.
+    super();
 
     // This collection holds the child layer elements.
     // @type {ol.Collection.<HTMLMapLayerBase>}
-    this.childLayerElementsCollection_ = this.constructor.setupChildLayerElementsObserver(this, this.layer.getLayers());
-  } // constructor
+    this.childLayerElementsCollection_ = this.constructor.getLiveChildLayerElementCollection(this, this.layer.getLayers());
+  }
 
   /**
    * Getters and Setters (for properties).
    */
+
+  /**
+   * @readonly
+   * @property {Array.<HTMLMapLayerBase>} layerElements
+   */
+  get layerElements () {
+    return this.childLayerElementsCollection_.getArray();
+  }
 
   /**
    * Customized public/private methods.
@@ -94,3 +133,5 @@ export default class HTMLMapLayerGroup extends HTMLMapLayerBase {
   }
 
 } // HTMLMapLayerGroup
+
+customElements.define(elementName, HTMLMapLayerGroup);