+++ src/elements/map-view/config.js
@@ -1,0 +1,3 @@
+export const elementName = "map-view";
+export const defaultMapType = "";
+export const defaultViewProjection = 'EPSG:3857';