+++ src/web-gis-components-lite.js
@@ -1,0 +1,15 @@
+/**
+ * The lite version doesn't include the web components polyfills.
+ */
+
+import './elements/map-view';
+
+import './elements/map-layer-group';
+import './elements/map-layer-twms';
+import './elements/map-layer-xyz';
+import './elements/map-layer-geojson';
+
+import './elements/map-control-defaults';
+import './elements/map-control-simple-layer-list';
+
+import './elements/map-interaction-defaults';