+++ src/web-gis-components.js
@@ -1,0 +1,8 @@
+/**
+ * The standard version comes with web components polyfills.
+ */
+
+import './third-party/webcomponentsjs/webcomponents-lite.js';
+import './third-party/webcomponentsjs/custom-elements-es5-adapter.js';
+
+import './web-gis-components-lite';