+++ History.md
@@ -1,5 +1,15 @@
 ## v.NEXT
 
+## v0.2.0
+
+- Replace "babel-cli + browserify" with webpack.
+- Include `web components` polyfills in bundle.
+- Build another bundle which does not include the polyfills.
+- Load openlayers from npm. Update it from v4.0.1 to v4.1.0.
+- Basic support for custom control elements.
+- Basic support for custom interaction elements.
+- Implement naive custom element for adding simple layer list onto the map.
+
 ## v0.1.0
 
 - Features: